+++ AlchemistCode/Cards/Uncommon/Knitbone.cs
@@ -13,7 +13,7 @@ public class Knitbone : AlchemistCard
 
     public Knitbone() : base(1, CardType.Skill, CardRarity.Uncommon, TargetType.Self)
     {
-        WithCalculatedBlock(5, 2, static (card, _) => Mixing.PlayedThisCombat(card.Owner), ValueProp.Move, 2);
+        WithCalculatedBlock(5, 1, static (card, _) => Mixing.PlayedThisCombat(card.Owner), ValueProp.Move, 0, 1);
         WithTips(_ => Mixing.MixRefTips());
     }